implementation exceptions

+++ ProxibanqueV4/dao/src/main/java/com/huios/dao/CompteRepository.java
@@ -17,7 +17,7 @@ public interface CompteRepository extends JpaRepository<Compte, Integer> {
 	
 	@Modifying
 	@Query("UPDATE Compte c SET c.numCompte = ?1, c.solde= ?2, c.dateOuverture = ?3 WHERE c.id= ?4")
-	public void modifierCompte(long numCompte, int solde, String dateOuverture, int idCompte);
+	public void modifierCompte(long numCompte, double solde, String dateOuverture, int idCompte);
 
 	@Query("FROM Compte a WHERE a.clientProprietaire.id = ?1 AND a.decouvert = ?2")
 	public CompteCourant trouverCompteCourant(int idclient, double decouvert);
